Fix bug in `find_index()`
Sometimes you can get an index of 0.  However, "if ($index)" returns false in
Perl if the index is 0.  Instead, we check the index var for
definedness.

This was causing us to build graphs that did not collect all of the
neighbors of a given node correctly, iff that node's index into the
graph array was 0.

Because we were building the graph data structure incorrectly, we were
failing to find paths when visual inspection tells us there are
definitely paths!