Update the internal graph representation
Specifically, we now use a Hash of ArrayRefs.

Profiling showed that the calls to O(n) `_find_index()`, which was used
all over the library was causing significant slowness.

Most methods now access the edges and vertices through getter methods
defined elsewhere in the library, rather than by reaching in and munging
things directly.  There are still a few that do this reaching in, but
they are somewhat lower-level, and will hopefully be refactored.

To make object copying work as expected in this new world, we did need
to add a dependency on Storable, which luckily is in core.  To make
object equality checking work, we needed to make `get_vertices()` and
`get_edges()` sort their output.  This (probably Quicksort, and thus O(n
log n)) cost was considered acceptable (for now) to get the
existing (admittedly incomplete) test suite passing.

Finally, as part of this work, we did need to edit *one* test, which was
relying on the array returned by `get_edges()` to be sorted in a
particular order.