Revert some centralized vertex definedness checks
This reverts most of commit e0a054c638f5ee554141fad8a9fd575e8c9fb587.

The implementation was bad and built a giant array and then grepped it,
which was very slow.  I don't think it's too much to ask for the
definedness check for users of the internal API (also known as: just me
^_^).