Revert the slow 'is_weighted' implementation
As of right now, there is little need for this check.

Further, it was implemented as a very slow nested loop.

If it were to be re-added in future, it should be a property of the
graph object at initialization, so the check can be simple and fast.

Obviously this API is still very much in flux!