Add a first draft of the WORDS example from SGB
SGB a.k.a. The Stanford GraphBase

Note that this version is horribly slow due to the inefficient
character-at-a-time word comparison in the `differ_by_one_char()`
subroutine.

After looking at the SGB version it's clear we need to change this to
use hashtables that store each variant of the word with a "hole" in it,
e.g., "words" has the variants "_ords", "w_rds", "wo_ds", "wor_s", and
"word_".