'exhaustive_search' sub takes additional arguments
Specifically, it can now look at the "path so far".  This will allow us
to update it to actually do something with that information in future --
at least that is the plan!  This required exposing the "path so far" to
"normal code" outside of debug mode.

Also, added a bit more method documentation.